App Asks for Location, and can detect if it is not turned on
So, in the main check for it now, the app will detect if Wifi is enabled, and grab GPS from that.
However, I will need to figure out a way to call the API call depending on if we have Zip Code or GPS. So far, I have a boolean, and will probably just check the boolean for what to do after that.

I will need to rewrite the Weather section a bit so I can have a function that receives either the Zip or the GPS coords, and then handle things from there.